# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the document for todos API",
    "version": "1.0.0",
    "title": "Portfolio API",
    "contact": {
      "email": "innocentingabire01@gmail.com"
    }
  },
  "tags": [
    {
      "name": "todo",
      "description": "My todos"
    },
    {
      "name": "user",
      "description": "The owner of todos"
    }
  ],
  "schemes": ["https", "http"],
  "paths": {
    "/api/todos": {
      "get": {
        "tags": ["todo"],
        "summary": "Get all todos associated with logged in user",
        "description": "",
        "operationId": "getTodos",
        "consumes": ["application/json"],
        "produces": ["application/json"],

        "responses": {
          "200": {
            "description": "request was successful"
          }
        }
      }
    },
    "/api/todos/{todoId}": {
      "get": {
        "tags": ["todo"],
        "summary": "Get one Todo",
        "description": "",
        "operationId": "getOneTodo",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "todoId",
            "description": "Getting one Todo post",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
    "/api/user/todos": {
      "post": {
        "tags": ["todo"],
        "summary": "Create a Todo",
        "description": "",
        "operationId": "createTodo",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "token"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Creating a Todo",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Todo"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Todo created successfully"
          },
          "401": {
            "description": "User not signed in"
          },
          "403": {
            "description": "Unauthorized access"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/todos/{todoId}": {
      "put": {
        "tags": ["todo"],
        "summary": "Update an existing todo",
        "description": "",
        "operationId": "updateTodo",
        "consumes": ["application/json"],
        "produces": ["application/xml"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "token"
          },
          {
            "in": "path",
            "name": "todoId",

            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Edit a blog post",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Todo"
            }
          }
        ],
        "responses": {
          "200":{
            "description":"sucess"
          },
          "400": {
            "description": "Invalid Input"
          },
          "404": {
            "description": "Todo not found"
          },
          "403": {
            "description": "Unauthorized access"
          }
        }
      },
      "delete": {
        "tags": ["todo"],
        "summary": "Delete an existing todo",
        "description": "",
        "operationId": "deleteTodo",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "token"
          },
          {
            "in": "path",
            "name": "blogId",
            "required": true
          }
        ],
        "responses": {
          "200":{
            "description":"success"
          },
          "400": {
            "description": "Invalid inputs"
          },
          "404": {
            "description": "Todo not found"
          },
          "403": {
            "description": "Unauthorized access"
          }
        }
      }
    },

    "/api/signin": {
      "post": {
        "tags": ["user"],
        "summary": "login",
        "description": "",
        "operationId": "loginUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "You need to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logged in Successfully"
          }
        }
      }
    }
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "properties": {
        "user":{
          "type":"string"
        },
        "title": {
          "type": "string"
        },
        "description":{
          "type":"string"
        },
        "priority": {
          "type": "string",
          "enum":["HIGH","MEDIUM","LOW"]
        }
      }
    },
    
    
    "User": {
      "type": "object",
      "properties": {
        "name":{
          "type":"string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}